pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file2ai"
version = "1.0.1"
description = "A comprehensive tool for file conversion, text extraction, and AI processing"
requires-python = ">=3.8"
authors = [
    { name = "Michael Jabbour", email = "michael@jabbour.dev" }
]
license = { file = "LICENSE" }
keywords = ["git", "export", "text"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "gitpython>=3.1.40",  # For git repository operations
    "python-docx>=0.8.11",  # For Word document support
    "openpyxl>=3.1.2",  # For Excel document support
    "weasyprint>=60.1",  # For HTML/PDF conversion
    "beautifulsoup4>=4.12.2",  # For HTML parsing
    "fpdf>=1.7.2",  # For PDF generation
    "python-pptx>=0.6.21",  # For PowerPoint support
    "requests>=2.31.0",  # For HTTP operations
    "pillow>=10.1.0",  # For image processing
    "pymupdf>=1.23.7",  # For PDF-to-image conversion
    "tqdm>=4.66.1"  # For progress bars
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",  # For testing
    "pytest-cov>=4.0",  # For test coverage
    "pytest-mock"  # For mocking in tests
]
web = [
    "flask>=2.0.0",  # For web interface
]

[project.urls]
"Source Code" = "https://github.com/michaeljabbour/file2ai"

# Document conversion is handled by pure Python packages:
# - python-docx: Word documents
# - openpyxl: Excel files
# - python-pptx: PowerPoint files
# - weasyprint: HTML/PDF conversion

[tool.setuptools]
py-modules = ["file2ai"]  # Single-file module

[tool.setuptools.packages.find]
exclude = ["exports*", "logs*", "tests*", "venv*", "*.egg-info"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = "-v --cov=. --cov-report=term-missing"
testpaths = ["."]
norecursedirs = ["exports", "logs", "venv", "__pycache__", "tests/backup"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
]
